Rename AusGlobeViewer->TerriaViewer, let terrain be specified.
